backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# AI框架
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1

# 数据处理
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 语音服务
edge-tts==6.1.9
pydub==0.25.1

# 系统控制
pyautogui==0.9.54
pywinauto==0.6.8
psutil==5.9.6
pywin32==306

# 图像处理（用于截图）
Pillow==10.1.0

# 剪贴板（用于文本处理）
pyperclip==1.8.2

# 工具库
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3
loguru==0.7.2
